+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Revision history for aws-arn
 
+## 0.2.0.0 -- 2022-??-??
+
+* `Network.AWS.ARN.colons`: Provide a `Lens'` instead of an unlawful `Iso'`
+* `Network.AWS.ARN.slashes`: Provide a `Lens'` instead of an unlawful `Iso'`
+
 ## 0.1.0.1 -- 2021-12-16
 
 * Support GHC 9.0.1, GHC 9.2.1, and `hashable <1.5`.